+++ ruby/lib/gitaly_server/repository_service.rb
@@ -8,7 +8,7 @@ module GitalyServer
       bridge_exceptions do
         repo_path = GitalyServer.repo_path(call)
 
-        Gitlab::Git::Repository.create(repo_path, bare: true, symlink_hooks_to: Gitlab.config.gitlab_shell.hooks_path)
+        Gitlab::Git::Repository.create(repo_path)
 
         Gitaly::CreateRepositoryResponse.new
       end