+++ ruby/lib/gitlab/git/repository.rb
@@ -2,14 +2,53 @@ module Gitlab
   module Git
     # These are monkey patches on top of the vendored version of Repository.
     class Repository
-      def self.from_gitaly(gitaly_repository, call)
-        new(
-          gitaly_repository,
-          GitalyServer.repo_path(call),
-          GitalyServer.gl_repository(call),
-          Gitlab::Git::GitlabProjects.from_gitaly(gitaly_repository, call),
-          GitalyServer.repo_alt_dirs(call)
-        )
+      class << self
+        def from_gitaly(gitaly_repository, call)
+          new(
+            gitaly_repository,
+            GitalyServer.repo_path(call),
+            GitalyServer.gl_repository(call),
+            Gitlab::Git::GitlabProjects.from_gitaly(gitaly_repository, call),
+            GitalyServer.repo_alt_dirs(call)
+          )
+        end
+
+        def create(repo_path)
+          FileUtils.mkdir_p(repo_path, mode: 0770)
+
+          # Equivalent to `git --git-path=#{repo_path} init [--bare]`
+          repo = Rugged::Repository.init_at(repo_path, true)
+          repo.close
+
+          symlink_hooks_to = Gitlab.config.gitlab_shell.hooks_path
+          create_hooks(repo_path, symlink_hooks_to) if symlink_hooks_to.present?
+        end
+
+        def create_hooks(repo_path, global_hooks_path)
+          local_hooks_path = File.join(repo_path, 'hooks')
+          real_local_hooks_path = :not_found
+
+          begin
+            real_local_hooks_path = File.realpath(local_hooks_path)
+          rescue Errno::ENOENT
+            # real_local_hooks_path == :not_found
+          end
+
+          # Do nothing if hooks already exist
+          unless real_local_hooks_path == File.realpath(global_hooks_path)
+            if File.exist?(local_hooks_path)
+              # Move the existing hooks somewhere safe
+              FileUtils.mv(
+                local_hooks_path,
+                "#{local_hooks_path}.old.#{Time.now.to_i}")
+            end
+
+            # Create the hooks symlink
+            FileUtils.ln_sf(global_hooks_path, local_hooks_path)
+          end
+
+          true
+        end
       end
 
       attr_reader :path